test(input): migrate fill
