+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-outline-label-floating-md-rtl-Mobile-Chrome-linux.png
