+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-shaped-outline-md-ltr-Mobile-Safari-linux.png
