+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-shaped-outline-md-rtl-Mobile-Safari-linux.png
