+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-shaped-solid-custom-md-ltr-Mobile-Safari-linux.png
