+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-shaped-solid-md-ltr-Mobile-Chrome-linux.png
