+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-solid-label-floating-md-ltr-Mobile-Chrome-linux.png
