+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-solid-label-floating-md-rtl-Mobile-Firefox-linux.png
