+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-solid-md-ltr-Mobile-Safari-linux.png
