+++ core/src/components/input/test/fill/input.e2e.ts-snapshots/input-fill-solid-md-rtl-Mobile-Safari-linux.png
